WAITP-1351 Move login token handling to prevent login modal displaying
